+++ src/routes/vigilants/+page.svelte
@@ -20,6 +20,7 @@
 	import { appState } from '$lib/models/appState';
 
 	import { routeTo } from '$lib/util';
+	import { showActionWillDeleteAssignment } from '../actionWillDeleteAssignment';
 
 	const modalStore = getModalStore();
 	const vigilantsStore = getAllVigilants();
@@ -32,8 +33,9 @@
 	$: disableDelete = $selected.length === 0;
 	const isAllSelected = handler.isAllSelected();
 
-	function deleteSelection() {
+	async function deleteSelection() {
 		if (disableDelete) return;
+		if (!(await showActionWillDeleteAssignment(modalStore))) return;
 		modalStore.trigger({
 			type: 'confirm',
 			title: m.delete_vigilant_modal_title(),
